Portfolio Website
This is a portfolio project built with HTML, CSS and javascript.

Within this project, the google analytics manager was implemented, a fixed scroll-back-to-top button, a fixed navigation bar, for good user experirence.

Additionally, a mailTo link for easy access to reach me and my CV in downloadable pdf format.

Likewise, links to my social media pages, including gitHub was included.

This project is at its simplest form for now, but will be built upon.